pyproject.toml
[project]
name = "sanictest"
version = "0.1.0"
description = ""
authors = [
    {name = "Станислав Рубцов",email = "frex@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sanic (>=25.3.0,<26.0.0)",
    "sanic-ext (>=24.12.0,<25.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

